Fixed login and sign up problems

+++ src/main/java/com/patientlink/backend/Authentication/JwtUtil.java
@@ -18,7 +18,7 @@ public class JwtUtil {
     @Value("${jwt.secret}")
     private String jwtSecret;
     @Value("${jwt.expiration}")
-    private int jwtExpiration;
+    private long jwtExpiration;
     
     private SecretKey key;
